# Improve invariants and performance in Chain