I realized that if we keep a suptype of Chain to note that items are non-empty, we can (at no cost, I think) simplify the code and make it safer.

There is a minor bit of ugliness in Append because to make it binary compatible, we need to put back some methods that were there before. I went ahead and did this to satisfy mima even though these classes are private to data.

Side note: unless we really need to expose classes like this, I'm a big fan of keeping them totally private.

also is a possible replacement for #3373