@johnynek 2020-07-27
I also avoid writing `case Empty =>` since that triggers a call to `def equals` which triggers a call to `iterator`. I managed to hit an infinite loop in one version of the code, but no reason to hit that for a sealed private class, we know Empty by exclusion and it is the least common case I would expect.