@johnynek 2020-07-27
While looking at this I found a performance improvement by not using an internal ArrayBuffer of 16% for large Chains. This is for iterator, which is used broadly. foldLeft exercises iterator.
```
This change:

[info] Benchmark                           Mode  Cnt         Score        Error  Units
[info] ChainBench.foldLeftLargeCatenable  thrpt    5        56.035 ±      0.661  ops/s
[info] ChainBench.foldLeftLargeChain      thrpt    5       199.145 ±      1.080  ops/s
[info] ChainBench.foldLeftLargeList       thrpt    5       158.537 ±      7.086  ops/s
[info] ChainBench.foldLeftLargeOldChain   thrpt    5       148.596 ±      1.560  ops/s
[info] ChainBench.foldLeftLargeVector     thrpt    5        77.262 ±      0.680  ops/s
[info] ChainBench.foldLeftSmallCatenable  thrpt    5  11846407.099 ±  22880.990  ops/s
[info] ChainBench.foldLeftSmallChain      thrpt    5  48157018.969 ± 246883.975  ops/s
[info] ChainBench.foldLeftSmallList       thrpt    5  55998179.055 ± 241976.102  ops/s
[info] ChainBench.foldLeftSmallOldChain   thrpt    5  55485668.350 ±  90607.034  ops/s
[info] ChainBench.foldLeftSmallVector     thrpt    5  14805536.562 ± 749006.508  ops/s


master:
[info] Benchmark                           Mode  Cnt         Score        Error  Units
[info] ChainBench.foldLeftLargeCatenable  thrpt    5        57.285 ±      1.891  ops/s
[info] ChainBench.foldLeftLargeChain      thrpt    5       170.942 ±      2.281  ops/s
[info] ChainBench.foldLeftLargeList       thrpt    5       159.709 ±      4.833  ops/s
[info] ChainBench.foldLeftLargeOldChain   thrpt    5       145.871 ±      7.981  ops/s
[info] ChainBench.foldLeftLargeVector     thrpt    5        76.742 ±      1.634  ops/s
[info] ChainBench.foldLeftSmallCatenable  thrpt    5  11798663.388 ± 606629.280  ops/s
[info] ChainBench.foldLeftSmallChain      thrpt    5  47750078.331 ± 473361.638  ops/s
[info] ChainBench.foldLeftSmallList       thrpt    5  56818234.798 ± 420088.798  ops/s
[info] ChainBench.foldLeftSmallOldChain   thrpt    5  63018070.265 ± 515461.221  ops/s
[info] ChainBench.foldLeftSmallVector     thrpt    5  14840988.935 ± 381582.742  ops/s
```

@LukaJCB can you look again?