@johnynek 2020-08-03
@barambani could you take another look (merged with master, a few minor cleanups).