@barambani 2020-08-03
LGTM on green. I think the failures are due to the update to scala 2.13.3. There are few missing parentheses on 'next' calls that now makes the build fail.