@johnynek 2020-08-03
Thanks @barambani I think I fixed it. Sorry for so many revisions here. :/ one more :+1: